Change timeout in_exp_comm to 10sec
The timeout was originally 2sec. An error which is now fixed sometimes caused devices to not be finished. Now, this eror should not occur anymore. However, if the error arises again, we have added an errorstatement (Commit: ecf50b87418726259ee75e9a09a0b798730fd5bd) and enlarge the timeout now, such that is is noticable but does not crash blacs